{{cookiecutter.project_slug}}/backend/app/core/security.py
from datetime import datetime, timedelta, timezone
import secrets

import jwt
from passlib.context import CryptContext

from app.core.config import settings

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

ALGORITHM = "HS256"


def create_access_token(subject: str, expires_delta: timedelta) -> str:
    expire = datetime.now(timezone.utc) + expires_delta
    to_encode = {"exp": expire, "sub": str(subject)}
    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt


def verify_password(plain_password: str, hashed_password: str) -> bool:
    return pwd_context.verify(plain_password, hashed_password)


def get_password_hash(password: str) -> str:
    return pwd_context.hash(password)

def generate_password_reset_token(email: str) -> str:
    return secrets.token_urlsafe(32)

def get_password_reset_token_expire_time() -> datetime:
    return datetime.now(timezone.utc) + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)

def is_password_reset_token_expired(expires_at: datetime) -> bool:
    return datetime.now(timezone.utc) > expires_at.replace(tzinfo=timezone.utc)
