{{cookiecutter.project_slug}}/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API para autenticación y gestión de usuarios con FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "annotated-types==0.7.0",
  "anyio==4.5.2",
  "bcrypt==4.0.1",
  "certifi==2025.4.26",
  "click==8.1.8",
  "dnspython==2.6.1",
  "email-validator==2.2.0",
  "exceptiongroup==1.2.2",
  "fastapi==0.115.12",
  "fastapi-cli==0.0.7",
  "greenlet==3.1.1",
  "h11==0.16.0",
  "httpcore==1.0.9",
  "httptools==0.6.4",
  "httpx==0.28.1",
  "idna==3.10",
  "jinja2==3.1.6",
  "markdown-it-py==3.0.0",
  "markupsafe==2.1.5",
  "mdurl==0.1.2",
  "passlib==1.7.4",
  "pydantic==2.10.6",
  "pydantic-core==2.27.2",
  "pydantic-settings==2.8.1",
  "pygments==2.19.1",
  "pyjwt==2.9.0",
  "python-dotenv==1.0.1",
  "python-multipart==0.0.20",
  "pyyaml==6.0.2",
  "rich==14.0.0",
  "rich-toolkit==0.14.5",
  "shellingham==1.5.4",
  "sniffio==1.3.1",
  "sqlalchemy==2.0.40",
  "sqlmodel==0.0.24",
  "starlette==0.44.0",
  "typer==0.15.3",
  "typing-extensions==4.13.2",
  "uvicorn==0.33.0",
  "uvloop==0.21.0",
  "watchfiles==0.24.0",
  "websockets==13.1",
  "psycopg",
  "ruff"
]

[tool.setuptools]
packages = ["app"]

[tool.uv]
dev-dependencies = [
  "pytest",
  "pytest-cov",
  "mypy",
  "pre-commit",
  "black",
  "isort",
  "alembic"
]